% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_univarie.R
\name{desc_quali}
\alias{desc_quali}
\title{desc_quali}
\usage{
desc_quali(
  data,
  ...,
  eff = TRUE,
  freq = TRUE,
  nb = 1,
  cum_freq = FALSE,
  NR = FALSE,
  tot = TRUE,
  pond = NULL,
  norm_pond = FALSE
)
}
\arguments{
\item{data}{base de données}

\item{...}{variables souhaitées}

\item{eff}{TRUE par défaut. Fait apparaître les effectifs par catégorie.}

\item{freq}{TRUE par défaut. idem mais avec les fréquences.}

\item{nb}{1 par défaut. Nombre de décimales pour les pourcentages.}

\item{cum_freq}{FALSE par défaut. Fréquences cumulées (veillez bien à ranger les modalités dans l'ordre souhaité)}

\item{NR}{FALSE par défaut. Les non-réponses comptent-elles parmi les catégories ?}

\item{tot}{TRUE par défaut. Ligne de total à la fin de chaque variable.}

\item{pond}{NULL par défaut. Vecteur contenant les poids.}

\item{norm_pond}{FALSE par défaut. Normalise les poids (poids moyen = 1).}
}
\value{
un tableau avec les effectifs et fréquences de chaque modalité de chaque variable sélectionnée
}
\description{
Permet d'obtenir un tri à plat de plusieurs variables catégorielles en même temps
}
\author{
Kanto Fiaferana & Julio Ricardo Davalos
}
