% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclure_agd.r
\name{exclure_agd}
\alias{exclure_agd}
\title{exclure_agd}
\usage{
exclure_agd(agd, na = TRUE, modalites = NULL)
}
\arguments{
\item{agd}{objet AGD (MCA, MFA, PCA...)}

\item{na}{indexe tous les NA (par défaut, TRUE)}

\item{modalites}{NULL par défaut. Sinon, liste de vecteurs nommés du nom de la variable et contenant la ou les modalités considérées}
}
\value{
vecteur ou liste des indexes, selon l'objet de départ
}
\description{
Permet d'obtenir l'objet regroupant les indexes des modalités à exclure des différents objets d'analyse géométrique des données. Par défaut, on exclue les non-réponses, mais on peut ajouter d'autres cas.
}
\author{
Julio Ricardo Davalos
}
