% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_agd.r
\name{format_agd}
\alias{format_agd}
\title{format_agd}
\usage{
format_agd(data, ttes_var = FALSE)
}
\arguments{
\item{data}{la base de données}

\item{ttes_var}{FALSE par défaut. On collerait alors tous les noms de variables à leurs modalités}
}
\value{
Un data.frame de même nature que la base de données utilisée.
}
\description{
Transforme une base de données préparée pour une analyse géométrique des données afin de supprimer les doublons dans la représentation graphiques. On colle les noms de variables aux modalités si ces dernières sont redondantes avec celles d'autres variables.
}
\author{
Julio Ricardo Davalos
}
