% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separe_ind_afm.R
\name{separe_ind_afm}
\alias{separe_ind_afm}
\title{separe_ind_afm}
\usage{
separe_ind_afm(afm, nb = 3, act_uniq = FALSE)
}
\arguments{
\item{afm}{objet MFA}

\item{nb}{nombre de chiffres à conserver pour toutes les statistiques produites.}

\item{act_uniq}{FALSE par défaut. Doit- extraire seulement les analyses factorielles actives ?}
}
\value{
Un data.frame comprenant le nom des variables, modalités, si elles sont actives ou supplémentaires ainsi que l'ensemble des données générées par les AGD de l'AFM (coordonnées, cos2, contributiosn etc.).
}
\description{
Permet d'extraire la liste des analyses séparées des individus d'une analyse factorielle multiple.
}
\author{
Julio Ricardo Davalos
}
