% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separe_quali_afm.R
\name{separe_quali_afm}
\alias{separe_quali_afm}
\title{separe_quali_afm}
\usage{
separe_quali_afm(afm, sup_dbl = TRUE, nb = Inf, act_uniq = FALSE)
}
\arguments{
\item{afm}{objet MFA}

\item{sup_dbl}{TRUE par défaut. La fonction colle le nom de la variable à la modalité en cas de doublon. Ici, on supprime cette concaténation pour plus de lisibilité dans le tableau final. Peut gêner la lecture dans le cas d'une représentation graphique basée uniquement sur les modalités.}

\item{nb}{nombre de chiffres à conserver pour toutes les statistiques produites.}

\item{act_uniq}{FALSE par défaut. Doit-on extraire seulement les analyses factorielles actives ?}
}
\value{
Un data.frame comprenant le nom des variables, modalités, si elles sont actives ou supplémentaires ainsi que l'ensemble des données générées par les AGD de l'AFM (coordonnées, cos2, contributiosn etc.).
}
\description{
Permet d'extraire les analyses séparées des variables d'une analyse factorielle multiple.
}
\author{
Julio Ricardo Davalos
}
