% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_ajust.R
\name{stat_ajust}
\alias{stat_ajust}
\title{stat_ajust}
\usage{
stat_ajust(..., ref = NULL)
}
\arguments{
\item{...}{le ou les modèles testés}

\item{ref}{NULL par défaut. Modèle saturé, afin que ses indicateurs AIC et BIC soient nuls.}
}
\value{
un tableau regroupant tous les indicateurs
}
\description{
Permet d'obtenir les indicateurs d'ajustement (variance, degrés de liberté, p-value, dissimilarité, AIC et BIC) pour les modèles log-linéaires pour lesquels les fonctions classiques ne fonctionnent pas. Merci à Maxime Parodi de nous avoir partagé son code.
}
