% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_quali_agd.r
\name{tab_quali_agd}
\alias{tab_quali_agd}
\title{tab_quali_agd}
\usage{
tab_quali_agd(agd, sup_dbl = TRUE, nb = 3)
}
\arguments{
\item{agd}{objet AGD (MCA, MFA, PCA...)}

\item{sup_dbl}{TRUE par défaut. La fonction colle le nom de la variable à la modalité en cas de doublon. Ici, on supprime cette concaténation pour plus de lisibilité dans le tableau final. Peut gêner la lecture dans le cas d'une représentation graphique basée uniquement sur les modalités.}

\item{nb}{nombre de chiffres à conserver pour toutes les statistiques produites.}
}
\value{
Un data.frame comprenant le nom des variables, modalités, si elles sont actives ou supplémentaires ainsi que l'ensemble des données générées par l'AGD (coordonnées, cos2, contributiosn etc.).
}
\description{
Permet d'obtenir un tableau contenant toutes les données de toutes les variables utilisées dans une analyse géométrique des données.
}
\author{
Julio Ricardo Davalos
}
