% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_quali.R
\name{desc_quali}
\alias{desc_quali}
\title{desc_quali (beta)}
\usage{
desc_quali(...)
}
\arguments{
\item{...}{variables souhaitées}
}
\value{
un tableau avec les effectifs et fréquences de chaque modalité de chaque variable
}
\description{
Permet d'obtenir une description de plusieurs variables en même temps avec la fonction freq()
}
\examples{
data <- data.frame(replicate(3,sample(0:3,1000,rep=TRUE)))
desc_quali(data$X1,data$X2,data$X3)
}
\author{
Kanto Fiaferana
}
