% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_quanti.R
\name{desc_quanti}
\alias{desc_quanti}
\title{desc_quanti}
\usage{
desc_quanti(
  data,
  ...,
  moy = TRUE,
  sd = TRUE,
  test.norm = TRUE,
  ic = TRUE,
  ic_seuil = 0.05,
  nb = 2,
  med = TRUE,
  quart = TRUE,
  minmax = TRUE,
  eff = TRUE,
  eff_na = FALSE,
  NR = TRUE
)
}
\arguments{
\item{data}{base de données}

\item{...}{variables souhaitées}

\item{moy}{TRUE par défaut. Moyenne de l'échantillon. Si FALSE, alors pas non plus d'écart-type.}

\item{sd}{TRUE par défaut. Ecart-type de l'échantillon}

\item{test.norm}{TRUE par défaut. P-value du test de normalité de l'échantillon (test de Shapiro-Wilk).}

\item{ic}{TRUE par défaut. Intervalle de confiance de la moyenne. N'apparait pas si moy = FALSE}

\item{ic_seuil}{risque de première espèce pour l'intervalle de confiance.}

\item{nb}{1 par défaut. Nombre de décimales.}

\item{med}{TRUE par défaut. Médiane de l'échantillon}

\item{quart}{TRUE par défaut. Quartiles Q1 et Q3 de l'échantillon.}

\item{minmax}{TRUE par défaut. Minimum et maximum .}

\item{eff}{TRUE par défaut. Effectifs.}

\item{eff_na}{FALSE par défaut. Remplace les effectifs par les effectifs avec les NA. N'apparait pas si eff = FALSE}

\item{NR}{TRUE par défaut. Supprime les non-réponses.}
}
\value{
un data.frame avec les indicateurs sélectionnés pour les variables quantitatives choisies
}
\description{
Permet d'obtenir les principaux indicateurs concernant un groupe de variables quantitatives
}
\author{
Julio Ricardo Davalos
}
