% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lprop_pctot.R
\name{lprop_pctot}
\alias{lprop_pctot}
\title{lprop_pctot}
\usage{
lprop_pctot(
  data,
  x,
  y,
  pond = NULL,
  tot_pond = FALSE,
  norm_pond = FALSE,
  num = TRUE,
  ch = 1,
  nr = c("no", "ifany", "always")
)
}
\arguments{
\item{data}{les données}

\item{x}{variable a mettre en ligne}

\item{y}{variable a mettre en colonne}

\item{pond}{variable de pondération (facultatif), doit absolument être écrit sous la forme data$pond}

\item{tot_pond}{logique. Si FALSE (valeur par deéfaut) alors les effectifs ne sont pas pondérés.}

\item{norm_pond}{TRUE par défaut. Si TRUE alors normalise les poids de la pondération (la somme des poids est alors égale a la somme des effectifs).}

\item{num}{logique. Si TRUE (valeur par défaut) alors les effectifs sont affiches}

\item{ch}{nombre de chiffres apres la virgule}

\item{nr}{comportements vis-a-vis des NA (voir table())}
}
\value{
un data.frame
}
\description{
Permet d'obtenir un tableau croise avec les pourcentages en lignes, mais ajoute une colonne avec le pourcentage total de la modalité.
En cas de variable de pondération, les pourcentages en lignes et de la modalité sont pondérés mais on peut choisir entre effectifs pondérés et effectifs réels.
}
\examples{
data <- data.frame(id=1:500, group = rep(LETTERS[1:2], 500/2), age = sample(18:30, 500, replace = TRUE), weights = abs(rnorm(500)))
lprop_pctot(data, age, group, data$weights)
}
\author{
Julio Ricardo Davalos
}
