% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lprop_pctot.R
\name{lprop_pctot}
\alias{lprop_pctot}
\title{lprop_pctot}
\usage{
lprop_pctot(x, m = TRUE, digits = 1)
}
\arguments{
\item{x}{un tableau}

\item{m}{logique. Si TRUE (valeur par défaut) alors les effectifs sont affichés}

\item{digits}{nombre de chiffres après la virgule}
}
\value{
un data.frame
}
\description{
Permet d'obtenir un tableau croisé avec les pourcentages en lignes, mais ajoute une colonne avec le pourcentage total de la modalité.
}
\examples{
lprop_pctot(as.table(rbind(c(1,2,3),c(4,5,6),c(7,8,9))))
}
\author{
Julio Ricardo Davalos
}
