% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_croise.R
\name{multi_croise}
\alias{multi_croise}
\title{multi_croise}
\usage{
multi_croise(
  data,
  var_princ,
  ...,
  NR = FALSE,
  pct_ligne = TRUE,
  nb = 1,
  pond = NULL,
  tot_pond = FALSE,
  norm_pond = FALSE
)
}
\arguments{
\item{data}{base de données}

\item{var_princ}{variable principale (en colonnes)}

\item{...}{variables à croiser avec la variable principale (en lignes)}

\item{NR}{FALSE par défaut. Compte les non-réponses.}

\item{pct_ligne}{TRUE par défaut. Pourcentages en ligne, sinon en colonne.}

\item{nb}{nombre de décimales pour les pourcentages. 1 par défaut}

\item{pond}{variable de pondération (facultatif), doit absolument être écrit sous la forme data$pond}

\item{tot_pond}{logique. Si FALSE (valeur par deéfaut) alors les effectifs ne sont pas pondérés.}

\item{norm_pond}{logique. Si TRUE alors normalise les poids de la pondération (la somme des poids est alors égale a la somme des effectifs). FALSE est la valeur par défaut.}
}
\value{
un tibble regroupant tous les tableaux croisés avec pourcentages et effectifs.
}
\description{
Permet d'obtenir un tableau croisant une variable d'intérêt avec une liste d'autres variables.
}
\author{
Julio Ricardo Davalos
}
