% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_quanti.R
\name{multi_quanti}
\alias{multi_quanti}
\title{multi_quanti}
\usage{
multi_quanti(
  data,
  var_princ,
  ...,
  moy = TRUE,
  anova = TRUE,
  sd = TRUE,
  ic = TRUE,
  ic_seuil = 0.05,
  nb = 2,
  med = TRUE,
  quant = 4,
  minmax = TRUE,
  eff = TRUE,
  eff_na = FALSE,
  NR = FALSE,
  msg = FALSE
)
}
\arguments{
\item{data}{base de données}

\item{var_princ}{variable principale (quantitative, en colonnes)}

\item{...}{variables catégorielles à croiser avec la variable principale (en lignes)}

\item{moy}{TRUE par défaut. Moyenne par modalité.}

\item{anova}{TRUE par défaut. Réalise un ANOVA et retourne sa p-value.}

\item{sd}{TRUE par défaut. Écart-type par modalité.}

\item{ic}{TRUE par défaut. Intervalle de confiance de la moyenne par modalité. N'apparait pas si moy = FALSE}

\item{ic_seuil}{risque de première espèce pour l'intervalle de confiance.}

\item{nb}{nombre de décimales pour la moyenne, l'écart-type et l'intervalle de confiance.}

\item{med}{TRUE par défaut. Médiane par modalité.}

\item{quant}{4 par défaut. Nombre de quantiles. Si la médiane est sélectionnée, elle sera ajoutée si besoin.}

\item{minmax}{TRUE par défaut. Minimum et maximum par modalité.}

\item{eff}{TRUE par défaut. Effectifs par modalité.}

\item{eff_na}{FALSE par défaut. Effectifs des non-réponses dans la variable quantitative par modalité.}

\item{NR}{FALSE par défaut. Garde les non-réponses des variables catégorielles.}

\item{msg}{FALSE par défaut. Envoie un message pour chaque variable terminée : utile si bug inexpliqué.}
}
\value{
Un tibble avec en colonne les indicateurs synthétiques de la variable d'intérêt selon les modalités des variables catégorielles choisies.
}
\description{
Permet d'obtenir un tableau détaillant une variable quantitative en fonction d'une liste de variables catégorielles.
}
\author{
Julio Ricardo Davalos
}
