% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_quanti.R
\name{multi_quanti}
\alias{multi_quanti}
\title{multi_quanti}
\usage{
multi_quanti(
  data,
  var_princ,
  ...,
  moy = TRUE,
  test.diffmoy = TRUE,
  force_anova = NULL,
  ic_test = 0.05,
  sd = TRUE,
  ic = TRUE,
  ic_seuil = 0.05,
  nb = 2,
  med = TRUE,
  quant = 4,
  minmax = TRUE,
  eff = TRUE,
  eff_na = FALSE,
  freq = TRUE,
  etoiles = FALSE,
  NR = FALSE,
  msg = FALSE
)
}
\arguments{
\item{data}{base de données}

\item{var_princ}{variable principale (quantitative, en colonnes)}

\item{...}{variables catégorielles à croiser avec la variable principale (en lignes)}

\item{moy}{TRUE par défaut. Moyenne par modalité.}

\item{test.diffmoy}{TRUE par défaut. Réalise un ANOVA et retourne la p-value du test d'égalité des moyennes adapté.
La procédure est la suivante: on réalise une ANOVA et un test de Shapiro-Wilk (test de normalité, remplacé par un test de Kolmogorov-Smirnov si l'effectif est supérieur à 5000 étant donné que le test de Shapiro-Wilk est sensible aux grands effectifs) sur celle-ci.
Si la normalité des résidus est rejetée alors le résultat final est un test de Kruskal-Wallis (on aura alors P(K)).
Si elle ne l'est pas alors on réalise un test de Levene (homoscédasticité).
Si l'homoscédasticité est rejetée alors le résultat final est un ANOVA de Welch (on aura alors P(W)), sinon on conserve l'ANOVA (on aura alors P(A)).}

\item{force_anova}{Permet de forcer la fonction à considérer la distribution comme normale. Vecteur des noms des variables catégorielles pour lesquelles forcer.
Peut être utile si on considère que la distribution est malgré tout suffisamment proche d'une distribution normale pour ne pas avoir d'incidence sur l'analyse.
Peut prendre la valeur TRUE si on veut forcer pour toutes les variables.}

\item{ic_test}{risque de première espèce des tests de d'égalité des moyennes.}

\item{sd}{TRUE par défaut. Écart-type par modalité.}

\item{ic}{TRUE par défaut. Intervalle de confiance de la moyenne par modalité. N'apparait pas si moy = FALSE}

\item{ic_seuil}{risque de première espèce pour l'intervalle de confiance.}

\item{nb}{nombre de décimales pour la moyenne, l'écart-type et l'intervalle de confiance.}

\item{med}{TRUE par défaut. Médiane par modalité.}

\item{quant}{4 par défaut. Nombre de quantiles. Si la médiane est sélectionnée, elle sera ajoutée si besoin.}

\item{minmax}{TRUE par défaut. Minimum et maximum par modalité.}

\item{eff}{TRUE par défaut. Effectifs par modalité.}

\item{eff_na}{FALSE par défaut. Effectifs des non-réponses dans la variable quantitative par modalité.}

\item{freq}{TRUE par défaut. Fréquence par modalité.}

\item{etoiles}{FALSE par défaut. Permet d'afficher les résultats des tests avec des étoiles à la place de la p.value en clair.}

\item{NR}{FALSE par défaut. Garde les non-réponses des variables catégorielles.}

\item{msg}{FALSE par défaut. Envoie un message pour chaque variable terminée : utile si bug inexpliqué.}
}
\value{
Un tibble avec en colonne les indicateurs synthétiques de la variable d'intérêt selon les modalités des variables catégorielles choisies.
}
\description{
Permet d'obtenir un tableau détaillant une variable quantitative en fonction d'une liste de variables catégorielles.
}
\author{
Julio Ricardo Davalos
}
