% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_quanti.R
\name{multi_quanti}
\alias{multi_quanti}
\title{multi_croise}
\usage{
multi_quanti(
  data,
  var_princ,
  ...,
  moy = TRUE,
  sd = TRUE,
  ic = TRUE,
  ic_seuil = 0.05,
  med = TRUE,
  quart = TRUE,
  minmax = TRUE,
  eff = TRUE,
  eff_na = FALSE,
  NR = TRUE
)
}
\arguments{
\item{data}{base de données}

\item{var_princ}{variable principale (quantitative, en colonnes)}

\item{...}{variables catégorielles à croiser avec la variable principale (en lignes)}

\item{moy}{TRUE par défaut. Moyenne par modalité.}

\item{sd}{TRUE par défaut. Écart-type par modalité.}

\item{ic}{TRUE par défaut. Intervalle de confiance de la moyenne par modalité.}

\item{ic_seuil}{risque de première espèce pour l'intervalle de confiance.}

\item{med}{TRUE par défaut. Médiane par modalité.}

\item{quart}{TRUE par défaut. Quartiles Q1 et Q3 par modalité.}

\item{minmax}{TRUE par défaut. Minimum et maximum par modalité.}

\item{eff}{TRUE par défaut. Effectifs par modalité.}

\item{eff_na}{FALSE par défaut. Effectifs par modalité en comptant les non-réponses à la place des effectifs.}

\item{NR}{TRUE par défaut. Supprime les modalités où il n'y a que des non-réponses.}
}
\value{
Un tabyl data.frame regroupant tous les tableaux croisés avec pourcentages et effectifs. Si les pourcentages sont en ligne et que les totaux sont activés alors la ligne de total est nommée "Ensemble" et la colonne de total est nommée "Total" et inversement pour les pourcentages en colonne.
}
\description{
Permet d'obtenir un tableau détaillant une variable quantitative en fonction d'une liste de variables catégorielles.
}
\author{
Julio Ricardo Davalos
}
