% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_ajust.R
\name{stat_ajust}
\alias{stat_ajust}
\title{stat_ajust}
\usage{
stat_ajust(..., ref = NULL)
}
\arguments{
\item{...}{le ou les modèles testés}

\item{ref}{NULL par défaut}
}
\value{
un tableau regroupant tous les indicateurs
}
\description{
Permet d'obtenir les indicateurs d'ajustement (variance, degrés de liberté, p-value, dissimilarité, AIC et BIC). Cela a un intérêt notamment concernant les modèles log-linéaires sur tableau de contingence pour lesquels la fonction glm n'arrive pas à produire d'AIC ou de BIC.
}
