% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_ind_agd.r
\name{tab_ind_agd}
\alias{tab_ind_agd}
\title{tab_ind_agd}
\usage{
tab_ind_agd(agd, nb = Inf, id = NULL)
}
\arguments{
\item{agd}{objet AGD (MCA, MFA, PCA...)}

\item{nb}{nombre de chiffres à conserver pour toutes les statistiques produites.}

\item{id}{vecteur contenant les identifiants des individus dans la base de données originale (facultatif). Doit être dans le même ordre que les individus de la base de données de l'AGD.}
}
\value{
Un data.frame comprenant les coordonnées, contribbutions, cos2 (et inertie intra individuelle pour les AFM) des individus et les modalités de ces individus
}
\description{
Permet d'obtenir un tableau contenant toutes les données des individus utilisés dans une analyse géométrique des données.
}
\author{
Julio Ricardo Davalos
}
